third_party/blink/web_tests/http/tests/security/cross-origin-window-event-exception-expected.txt

Cross-origin access to 'window.event' should throw a SecurityError.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS frame.contentWindow.event threw exception SecurityError: Failed to read a named property 'event' from 'Window': Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame..
PASS frame.contentWindow.event = 1; threw exception SecurityError: Failed to set a named property 'event' on 'Window': Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame..
PASS successfullyParsed is true

TEST COMPLETE
